+++ Cargo-minimal.lock
@@ -217,6 +217,7 @@ dependencies = [
  "arbitrary",
  "bincode",
  "bitcoin-consensus-encoding",
+ "bitcoin-crypto",
  "bitcoin-internals 0.5.0",
  "bitcoin-units 0.3.0",
  "bitcoin_hashes 0.20.0",
