+++ bitcoin/src/blockdata/script/mod.rs
@@ -59,7 +59,6 @@ use io::{BufRead, Write};
 
 use self::witness_version::WitnessVersion;
 use crate::consensus::{encode, Decodable, Encodable};
-use crate::internal_macros::impl_asref_push_bytes;
 use crate::key::WPubkeyHash;
 use crate::opcodes::all::*;
 use crate::opcodes::Opcode;
@@ -92,8 +91,6 @@ pub use self::error::{
     WitnessScriptSizeError,
 };
 pub(crate) use self::owned::ScriptBufExtPriv;
-
-impl_asref_push_bytes!(ScriptHash, WScriptHash);
 
 /// Constructs a new [`WitnessScriptBuf`] containing the script code used for spending a P2WPKH output.
 ///
@@ -251,10 +248,11 @@ pub mod error {
 
     #[rustfmt::skip]        // Keep public re-exports separate.
     #[doc(inline)]
-    pub use super::push_bytes::{PushBytesError, ScriptIntError};
+    pub use super::push_bytes::ScriptIntError;
     #[doc(no_inline)]
     pub use primitives::script::error::{
-        FromHexError, RedeemScriptSizeError, ScriptBufDecoderError, WitnessScriptSizeError,
+        FromHexError, PushBytesError, RedeemScriptSizeError, ScriptBufDecoderError,
+        WitnessScriptSizeError,
     };
 
     /// Ways that a script might fail. Not everything is split up as
