+++ primitives/Cargo.toml
@@ -15,13 +15,14 @@ exclude = ["api", "tests", "contrib"]
 
 [features]
 default = ["std", "hex"]
-std = ["alloc", "hashes/std", "hex-stable?/std", "hex-unstable?/std", "internals/std", "units/std"]
-alloc = ["hashes/alloc", "hex-stable?/alloc", "hex-unstable?/alloc", "internals/alloc", "units/alloc"]
-serde = ["dep:serde", "hashes/serde", "internals/serde", "units/serde", "alloc", "hex"]
-arbitrary = ["dep:arbitrary", "units/arbitrary"]
+std = ["alloc", "crypto/std", "hashes/std", "hex-stable?/std", "hex-unstable?/std", "internals/std", "units/std"]
+alloc = ["crypto/alloc", "hashes/alloc", "hex-stable?/alloc", "hex-unstable?/alloc", "internals/alloc", "units/alloc"]
+serde = ["dep:serde", "crypto/serde", "hashes/serde", "internals/serde", "units/serde", "alloc", "hex"]
+arbitrary = ["dep:arbitrary", "crypto/arbitrary", "units/arbitrary"]
 hex = ["dep:hex-stable", "dep:hex-unstable", "hashes/hex", "internals/hex"]
 
 [dependencies]
+crypto = { package = "bitcoin-crypto", path = "../crypto", version = "0.2.0", default-features = false }
 encoding = { package = "bitcoin-consensus-encoding", path = "../consensus_encoding", version = "0.2.0", default-features = false }
 hashes = { package = "bitcoin_hashes", path = "../hashes", version = "0.20.0", default-features = false }
 internals = { package = "bitcoin-internals", path = "../internals", version = "0.5.0" }
