+++ primitives/src/hash_types/witness_script_hash.rs
@@ -22,6 +22,7 @@ use crate::script::{WitnessScript, MAX_WITNESS_SCRIPT_SIZE};
 pub struct WScriptHash(sha256::Hash);
 
 super::impl_debug!(WScriptHash);
+crate::internal_macros::impl_asref_push_bytes!(WScriptHash);
 
 impl WScriptHash {
     /// Constructs a new `WScriptHash` after first checking the script size.
